requirements.txt
annotated-types==0.7.0
anyio==4.5.0
certifi==2024.8.30
distro==1.9.0
exceptiongroup==1.2.2
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jiter==0.5.0
MouseInfo==0.1.3
openai==1.46.1
pillow==10.4.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pydantic==2.9.2
pydantic_core==2.23.4
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Quartz==10.3.1
pyperclip==1.9.0
PyQt5==5.15.2
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
rubicon-objc==0.4.9
sniffio==1.3.1
SQLAlchemy==2.0.35
tqdm==4.66.5
typing_extensions==4.12.2
python-dotenv
pynput
pyautogui
